changed hierarchy of modules
