+++ ctapipe/instrument/__init__.py
@@ -1,2 +1,5 @@
 from .util_functions import *
 from .InstrumentDescription import *
+from .OpticsDescription import *
+from .CameraDescription import *
+from .TelescopeDescription import *
