--- ctapipe/instrument/telescope/__init__.py
@@ -1,1 +1,0 @@
-from .TelescopeDescription import *
