--- ctapipe/instrument/telescope/camera/__init__.py
@@ -1,1 +1,0 @@
-from .CameraDescription import *
