--- ctapipe/instrument/telescope/optics/__init__.py
@@ -1,1 +1,0 @@
-from .OpticsDescription import *
